Demonstrate serial version UIDs
