+++ src/application/Person.java
@@ -1,0 +1,25 @@
+package application;
+
+import java.io.Serializable;
+
+
+	public class Person implements Serializable {
+
+			private static final long serialVersionUID = 1L;
+			/* purpose of this is to prevent you from saving one
+			 * version of the class and then loading a different version
+			 */
+			private String name;
+			private int id;
+			
+			public Person(String name, int id) {
+				this.name = name;
+				this.id = id;
+			}
+
+			@Override
+			public String toString() {
+				return "Person [name=" + name + ", id=" + id + "]";
+			}
+		}
+
